Rename "new_quiz" references to "quiz"
