+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_button_replyed.png
