+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_displayTextResponse_received.png
