+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_displayText_received.png
