+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_empty_state.png
