+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_error_state_retry.png
