+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_horizontalButtons_replyed.png
