+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_multipleChoices_received.png
