+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_received_many_messages.png
