+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_scroll_to_end.png
