+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_showHelpTutorial_replyed.png
