+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_showStealthTutorial_replyed.png
