+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_singleChoice_replyed.png
