+++ test/app/features/quiz/presentation/quiz/goldens/ci/quiz_yesno_replyed.png
